requirements.txt
astronomer-cosmos
dbt-bigquery
apache-airflow-providers-slack
airflow-provider-hex